@@ -6,23 +6,20 @@
     "@fortawesome/fontawesome-svg-core": "^1.2.15",
     "@fortawesome/free-solid-svg-icons": "^5.7.2",
     "@fortawesome/react-fontawesome": "^0.1.4",
-    "@reduxjs/toolkit": "^1.8.0",
     "@sambego/storybook-state": "^1.3.2",
-    "@storybook/react": "^4.1.13",
+    "@storybook/react": "^6.4.19",
     "eslint-plugin-babel": "^5.3.1",
-    "final-form": "^4.20.6",
-    "firebase": "^8.10.1",
     "prop-types": "^15.7.2",
-    "react": "^16.8.3",
-    "react-dom": "^16.8.3",
-    "react-final-form": "^6.5.8",
-    "react-redux": "^7.2.6",
-    "react-router-dom": "^5.3.0",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "react-hook-form": "^7.28.0",
+    "react-query": "^3.34.16",
+    "react-router-dom": "^6.2.2",
     "react-scripts": "^5.0.0",
     "react-very-simple-alerts": "^1.0.1",
-    "redux": "^4.0.1",
-    "styled-components": "^4.1.3",
-    "uuid": "^3.3.2"
+    "styled-components": "^5.3.3",
+    "uuid": "^3.3.2",
+    "zustand": "^3.7.1"
   },
   "scripts": {
     "start": "ESLINT_NO_DEV_ERRORS='true' react-scripts start",
@@ -52,11 +49,9 @@
     "eslint-plugin-import": "^2.25.4",
     "eslint-plugin-jest": "^26.1.1",
     "eslint-plugin-jsx-a11y": "^6.5.1",
-    "eslint-plugin-prettier": "^3.0.1",
     "eslint-plugin-react": "^7.29.2",
     "eslint-plugin-react-hooks": "^4.3.0",
-    "firebase-tools": "^6.4.0",
-    "husky": "^1.3.1",
+    "husky": "^7.0.4",
     "prettier": "^2.5.1",
     "pretty-quick": "^1.10.0"
   },
