@@ -1,17 +1,14 @@
 import React from "react";
 import ReactDOM from "react-dom";
-
-import { Provider } from "react-redux";
-import store from "./store/store";
-
+import { BrowserRouter } from "react-router-dom";
 import App from "./App";
 import * as serviceWorker from "./serviceWorker";
 import "./fontawesome/initFaIcons";
 
 ReactDOM.render(
-  <Provider store={store}>
+  <BrowserRouter>
     <App />
-  </Provider>,
+  </BrowserRouter>,
   document.getElementById("root")
 );
 
